Updated Line 79 for error handling
Any ciphertext that is encrypted from plaintext with multiple lines, i.e.:

two
lines

would throw a UnicodeDecodeError when running the decode() on it. Updated code handles that issue and also replaces the characters to properly display on the website. The index.html change needs to update as well for this to properly function.